bar_scraper/requirements.txt
googlemaps==4.10.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
python-dotenv==1.0.1
pandas==2.2.2
